+++ app/models/EligibilityCalculatorResult.scala
@@ -16,4 +16,9 @@
 
 package models
 
-case class EligibilityCalculatorResult(messageKey: String, monetaryGain: Option[Double] = None, messageParam: Option[String] = None, messageParam2: Option[String] = None)
+case class EligibilityCalculatorResult(
+                                        messageKey: String,
+                                        monetaryGain: Option[BigDecimal] = None,
+                                        messageParam: Option[String] = None,
+                                        messageParam2: Option[String] = None
+                                      )
